# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "km_engine API",
    "description": "Knowledge Management Engine API (WP5, Gate G5)",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/entities": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Lista entit\u00e0 filtrate per visibilit\u00e0",
        "description": "Ottieni lista entit\u00e0 visibili all'utente autenticato.\n\n- **Filtro visibilit\u00e0**: applicato automaticamente in base a ruoli/teams\n- **Admin bypass**: gli admin vedono tutte le entit\u00e0\n- **lang**: se specificata e != \"en\", i contenuti non tradotti hanno flag untranslated",
        "operationId": "list_entities_api_v1_entities_get",
        "parameters": [
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro per label",
              "title": "Label"
            },
            "description": "Filtro per label"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro per tipo",
              "title": "Type"
            },
            "description": "Filtro per tipo"
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lingua risposta (override)",
              "title": "Lang"
            },
            "description": "Lingua risposta (override)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AuthSettings"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Settings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Entities Api V1 Entities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entities/{entity_id}": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Ottieni entit\u00e0 per ID",
        "description": "Ottieni un'entit\u00e0 specifica con i suoi fatti correnti.\n\n- **404**: se l'entit\u00e0 non esiste o non \u00e8 visibile\n- **Filtro visibilit\u00e0**: applicato su entit\u00e0 e fatti",
        "operationId": "get_entity_api_v1_entities__entity_id__get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lingua risposta (override)",
              "title": "Lang"
            },
            "description": "Lingua risposta (override)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AuthSettings"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Settings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Entity Api V1 Entities  Entity Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entities/{entity_id}/facts": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Ottieni fatti di un'entit\u00e0",
        "description": "Ottieni i fatti di un'entit\u00e0, con opzionale filtro temporale.\n\n- **at_time**: se specificato, restituisce i fatti validi a quel timestamp (FR5.3)\n- **Filtro visibilit\u00e0**: applicato a ogni fatto",
        "operationId": "list_facts_api_v1_entities__entity_id__facts_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "at_time",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Timestamp per query storica 'al tempo T'",
              "title": "At Time"
            },
            "description": "Timestamp per query storica 'al tempo T'"
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lingua risposta (override)",
              "title": "Lang"
            },
            "description": "Lingua risposta (override)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AuthSettings"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Settings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Facts Api V1 Entities  Entity Id  Facts Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/entities/{entity_id}/relations": {
      "get": {
        "tags": [
          "entities"
        ],
        "summary": "Ottieni relazioni di un'entit\u00e0",
        "description": "Ottieni le relazioni RELATES_TO di un'entit\u00e0.\n\n- **Filtro visibilit\u00e0**: applicato all'entit\u00e0 target di ogni relazione",
        "operationId": "list_relations_api_v1_entities__entity_id__relations_get",
        "parameters": [
          {
            "name": "entity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entity Id"
            }
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lingua risposta (override)",
              "title": "Lang"
            },
            "description": "Lingua risposta (override)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AuthSettings"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Settings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Relations Api V1 Entities  Entity Id  Relations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Ricerca full-text",
        "description": "Ricerca full-text su label e value dei fatti.\n\n- **Implementazione**: CONTAINS (documentato; per performance usare indici Neo4j full-text)\n- **Filtro visibilit\u00e0**: applicato a tutti i risultati\n- **match_type**: \"entity\" o \"fact\" indica il tipo di match",
        "operationId": "search_endpoint_api_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Testo da cercare",
              "title": "Q"
            },
            "description": "Testo da cercare"
          },
          {
            "name": "label",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtro per label",
              "title": "Label"
            },
            "description": "Filtro per label"
          },
          {
            "name": "lang",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Lingua risposta (override)",
              "title": "Lang"
            },
            "description": "Lingua risposta (override)"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AuthSettings"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Settings"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Search Endpoint Api V1 Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AuthSettings": {
        "properties": {
          "pg_dsn": {
            "type": "string",
            "title": "Pg Dsn",
            "default": "postgresql://km:km_dev_password@localhost:5432/km_engine"
          },
          "jwt_secret": {
            "type": "string",
            "title": "Jwt Secret",
            "default": "dev-only-secret-change-in-prod"
          },
          "jwt_algorithm": {
            "type": "string",
            "title": "Jwt Algorithm",
            "default": "HS256"
          },
          "access_token_ttl": {
            "type": "string",
            "format": "duration",
            "title": "Access Token Ttl",
            "default": "PT15M"
          },
          "refresh_token_ttl": {
            "type": "string",
            "format": "duration",
            "title": "Refresh Token Ttl",
            "default": "P14D"
          },
          "tenant": {
            "type": "string",
            "title": "Tenant",
            "default": "default"
          },
          "admin_username": {
            "type": "string",
            "title": "Admin Username",
            "default": "admin"
          },
          "admin_password": {
            "type": "string",
            "title": "Admin Password",
            "default": ""
          }
        },
        "type": "object",
        "title": "AuthSettings",
        "description": "Impostazioni auth: DSN Postgres, secret JWT, TTL dei token, bootstrap admin.\n\nLette da variabili d'ambiente con prefisso ``KM_`` (es. ``KM_PG_DSN``) e,\nin dev, dal file ``.env`` della repo root."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}